Game: (refactor) extracted classes and functions into different files
